Remove public
